Refactor position types to Q10.6 fixed point
Changed all position and velocity fields in game entities to use int16_t Q10.6 fixed point for improved precision and memory efficiency. Updated rendering and movement logic to support fixed point arithmetic, including new goToCoords function for terminal positioning. Adjusted bullet, asteroid, and UFO spawning and movement to use fixed point values, and improved rendering to handle erasing and drawing based on movement.

+++ Inc/game_master.h
@@ -1,6 +1,18 @@
 #ifndef _GAME_MASTER_H_
 #define _GAME_MASTER_H_
 
+
+	/*
+	 * Why use int16 for position?
+	 * We use a 10.6 fixed int16 for the x and y axis
+	 * This is to both save on data as the precesion of int32 is not needed.
+	 * with the 10.6 fixed point int16 the range is −512.0 to +511.984375.
+	 * Which is plenty as most screen terminals is ~200 cells wide, at fullscreen.
+	 *
+	 * We also allow for x and y to be negative only because we will clamp the values later in our code.
+	 * And having both velocity and the position in the same type makes calculations easier.
+	 */
+
 	/* ----- Dependencies ----- */
 	#include <stdint.h>
 	#include <stdlib.h>
@@ -20,32 +32,32 @@
 	} gameState_t;
 
 	typedef struct {
-		uint8_t x, y;
-		int32_t vX, vY; // 16.16
+		int16_t x, y; 	// Q10.6 Fixed point
+		int16_t vX, vY; // Q10.6 Fixed point
 		uint8_t lvl;
 		uint8_t powerUp;
 	} spaceship_t;
 
 	typedef struct {
 		uint8_t active;
-		uint8_t x, y;
-		int32_t vX, vY; // 16.16
+		int16_t x, y; 	// Q10.6 Fixed point
+		int16_t vX, vY; // Q10.6 Fixed point
 		uint8_t dmg;
 		uint8_t type;
 	} bullet_t;
 
 	typedef struct {
 		uint8_t active;
-		uint8_t x, y;
-		int32_t vX, vY; // 16.16
+		int16_t x, y; 	// Q10.6 Fixed point
+		int16_t vX, vY; // Q10.6 Fixed point
 		uint8_t lives;
 		uint8_t type;
 	} asteroid_t;
 
 	typedef struct {
 		uint8_t active;
-		uint8_t x, y;
-		int32_t vX, vY; // 16.16
+		int16_t x, y; 	// Q10.6 Fixed point
+		int16_t vX, vY; // Q10.6 Fixed point
 		uint8_t lives;
 		uint8_t type;
 		uint8_t shotDelay; // Number of ticks before alien shoots again
@@ -59,7 +71,7 @@
 	void addUfo();
 	void updateUfos();
 
-	void addBullet(uint8_t x, uint8_t y, uint32_t vY);
+	void addBullet(int16_t x, int16_t y, int16_t vY);
 	void updateBullets();
 
 #endif
